renamed, rebased Constants.java and added some properties
