--- src/main/java/executor/service/Constants.java
@@ -1,15 +1,0 @@
-package executor.service;
-
-/**
- * Class for application`s constants.
- * */
-public final class Constants {
-
-    private Constants() {
-        throw new AssertionError("non-instantiable class");
-    }
-
-    public static final String CORE_POOL_SIZE = "core.pool.size";
-    public static final String KEEP_ALIVE_TIME = "keep.alive.time";
-
-}
