+++ src/main/java/executor/service/config/properties/PropertiesConstants.java
@@ -1,0 +1,17 @@
+package executor.service.config.properties;
+
+/**
+ * Class for application`s constants.
+ * */
+public final class PropertiesConstants {
+
+    private PropertiesConstants() {
+        throw new AssertionError("non-instantiable class");
+    }
+
+    public static final String THREAD_POOL_PROPERTIES = "thread-pool.properties";
+    public static final String CORE_POOL_SIZE = "core.pool.size";
+    public static final String KEEP_ALIVE_TIME = "keep.alive.time";
+    public static final String COMMONS_CONFIGURATION_PROPERTIES = "commons-configuration.properties";
+
+}
